Refactoring the notification filter and the ticket generator
